Move build and serve commands into cli module
